fix: resolve theme flicker and make dashboard the root URL
Fix FOUC (Flash of Unstyled Content) where the blocking theme script
didn't match next-themes defaultTheme="dark" — first-time visitors with
a light system preference saw a white flash before dark mode applied.
Also use bg-background instead of bg-white dark:bg-background on <main>
to avoid any transient white flash during hydration.

Move dashboard from /dashboard to / so ganamos.earth serves the feed
directly. Old /dashboard URLs permanently redirect via next.config and
an in-app redirect page. All internal router.push, href, revalidatePath,
and middleware references updated across ~25 files.

Made-with: Cursor
